refactor(mm): split memory space modules
